Deprecate DataApiMixin.
