+++ docs/tutorials/crash-course/model-and-schema.md
@@ -21,7 +21,7 @@ Later you will also learn about the **editing UI**, but for now, all you need to
 
 ### Model
 
-The first and most important part of the editing engine is the model. The model is an HTML-like structure that represents the content of the editor. When the {@link module:core/editor/utils/dataapimixin~DataApi#setData `editor.setData()`} method is called, HTML passed as the argument is converted into the model. Then, when the {@link module:core/editor/utils/dataapimixin~DataApi#getData `model.getData()`} method is called, the model is converted back to HTML.
+The first and most important part of the editing engine is the model. The model is an HTML-like structure that represents the content of the editor. When the {@link module:core/editor/editor~Editor#setData `editor.setData()`} method is called, HTML passed as the argument is converted into the model. Then, when the {@link module:core/editor/editor~Editor#getData `model.getData()`} method is called, the model is converted back to HTML.
 
 One major difference between the model and HTML is that in the model, both text and elements can have attributes.
