+++ packages/ckeditor5-core/src/editor/editor.ts
@@ -464,6 +464,68 @@ export default abstract class Editor extends ObservableMixin() {
 	}
 
 	/**
+	 * Sets the data in the editor.
+	 *
+	 * ```ts
+	 * editor.setData( '<p>This is editor!</p>' );
+	 * ```
+	 *
+	 * If your editor implementation uses multiple roots, you should pass an object with keys corresponding
+	 * to the editor root names and values equal to the data that should be set in each root:
+	 *
+	 * ```ts
+	 * editor.setData( {
+	 *     header: '<p>Content for header part.</p>',
+	 *     content: '<p>Content for main part.</p>',
+	 *     footer: '<p>Content for footer part.</p>'
+	 * } );
+	 * ```
+	 *
+	 * By default the editor accepts HTML. This can be controlled by injecting a different data processor.
+	 * See the {@glink features/markdown Markdown output} guide for more details.
+	 *
+	 * @param data Input data.
+	 */
+	public setData( data: string | Record<string, string> ): void {
+		this.data.set( data );
+	}
+
+	/**
+	 * Gets the data from the editor.
+	 *
+	 * ```ts
+	 * editor.getData(); // -> '<p>This is editor!</p>'
+	 * ```
+	 *
+	 * If your editor implementation uses multiple roots, you should pass root name as one of the options:
+	 *
+	 * ```ts
+	 * editor.getData( { rootName: 'header' } ); // -> '<p>Content for header part.</p>'
+	 * ```
+	 *
+	 * By default, the editor outputs HTML. This can be controlled by injecting a different data processor.
+	 * See the {@glink features/markdown Markdown output} guide for more details.
+	 *
+	 * A warning is logged when you try to retrieve data for a detached root, as most probably this is a mistake. A detached root should
+	 * be treated like it is removed, and you should not save its data. Note, that the detached root data is always an empty string.
+	 *
+	 * @param options Additional configuration for the retrieved data.
+	 * Editor features may introduce more configuration options that can be set through this parameter.
+	 * @param options.rootName Root name. Defaults to `'main'`.
+	 * @param options.trim Whether returned data should be trimmed. This option is set to `'empty'` by default,
+	 * which means that whenever editor content is considered empty, an empty string is returned. To turn off trimming
+	 * use `'none'`. In such cases exact content will be returned (for example `'<p>&nbsp;</p>'` for an empty editor).
+	 * @returns Output data.
+	 */
+	public getData( options?: {
+		rootName?: string;
+		trim?: 'empty' | 'none';
+		[ key: string ]: unknown;
+	} ): string {
+		return this.data.get( options );
+	}
+
+	/**
 	 * Loads and initializes plugins specified in the configuration.
 	 *
 	 * @returns A promise which resolves once the initialization is completed, providing an array of loaded plugins.
