+++ packages/ckeditor5-core/src/editor/utils/dataapimixin.ts
@@ -9,35 +9,15 @@
 
 import type Editor from '../editor.js';
 
-import type { Constructor, Mixed } from '@ckeditor/ckeditor5-utils';
+import type { Constructor } from '@ckeditor/ckeditor5-utils';
 
 /**
  * Implementation of the {@link module:core/editor/utils/dataapimixin~DataApi}.
+ *
+ * @deprecated This functionality is already implemented by the `Editor` class.
  */
-export default function DataApiMixin<Base extends Constructor<Editor>>( base: Base ): Mixed<Base, DataApi> {
-	abstract class Mixin extends base implements DataApi {
-		public setData( data: string | Record<string, string> ): void {
-			this.data.set( data );
-		}
-
-		public getData( options?: {
-			rootName?: string;
-			trim?: 'empty' | 'none';
-			[ key: string ]: unknown;
-		} ): string {
-			return this.data.get( options );
-		}
-	}
-
-	return Mixin as any;
-}
-
-// Backward compatibility with `mix`.
-{
-	const mixin = ( DataApiMixin as any )( Object );
-
-	( DataApiMixin as any ).setData = mixin.prototype.setData;
-	( DataApiMixin as any ).getData = mixin.prototype.getData;
+export default function DataApiMixin<Base extends Constructor<Editor>>( base: Base ): Base {
+	return base;
 }
 
 /**
@@ -47,6 +27,8 @@ export default function DataApiMixin<Base extends Constructor<Editor>>( base: Ba
  * This interface is not a part of the {@link module:core/editor/editor~Editor} class because one may want to implement
  * an editor with multiple root elements, in which case the methods for setting and getting data will need to be implemented
  * differently.
+ *
+ * @deprecated This interface is implemented by all `Editor` instances by default.
  */
 export interface DataApi {
