+++ packages/ckeditor5-core/tests/editor/editor.js
@@ -17,6 +17,8 @@ import Command from '../../src/command.js';
 import EditingKeystrokeHandler from '../../src/editingkeystrokehandler.js';
 import { expectToThrowCKEditorError } from '@ckeditor/ckeditor5-utils/tests/_utils/utils.js';
 import CKEditorError from '@ckeditor/ckeditor5-utils/src/ckeditorerror.js';
+import testUtils from '../../tests/_utils/utils.js';
+import { getData, setData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model.js';
 
 class TestEditor extends Editor {
 	static create( config ) {
@@ -1248,6 +1250,69 @@ describe( 'Editor', () => {
 						editor.plugins.get( PluginF ).afterInit
 					);
 				} );
+		} );
+	} );
+
+	describe( 'data API', () => {
+		let editor;
+
+		beforeEach( () => {
+			// eslint-disable-next-line new-cap
+			class CustomEditor extends Editor {}
+
+			editor = new CustomEditor();
+			editor.model.document.createRoot( '$root', 'main' );
+			editor.model.document.createRoot( '$root', 'secondRoot' );
+			editor.model.schema.extend( '$text', { allowIn: '$root' } );
+			editor.fire( 'ready' ); // (#6139)
+		} );
+
+		afterEach( async () => {
+			await editor.destroy();
+		} );
+
+		describe( 'setData()', () => {
+			it( 'should be added to editor interface', () => {
+				expect( editor ).have.property( 'setData' ).to.be.a( 'function' );
+			} );
+
+			it( 'should set data of the first root', () => {
+				editor.setData( 'foo' );
+
+				expect( getData( editor.model, { rootName: 'main', withoutSelection: true } ) ).to.equal( 'foo' );
+			} );
+		} );
+
+		describe( 'getData()', () => {
+			testUtils.createSinonSandbox();
+
+			it( 'should be added to editor interface', () => {
+				expect( editor ).have.property( 'getData' ).to.be.a( 'function' );
+			} );
+
+			it( 'should get data of the first root', () => {
+				setData( editor.model, 'foo' );
+
+				expect( editor.getData() ).to.equal( 'foo' );
+			} );
+
+			it( 'should get data of the second root', () => {
+				setData( editor.model, 'bar', { rootName: 'secondRoot' } );
+
+				expect( editor.getData( { rootName: 'secondRoot' } ) ).to.equal( 'bar' );
+			} );
+
+			it( 'should pass options object to data.get() method internally', () => {
+				const spy = testUtils.sinon.spy( editor.data, 'get' );
+				const options = { rootName: 'main', trim: 'none' };
+
+				setData( editor.model, 'foo' );
+
+				expect( editor.getData( options ) ).to.equal( 'foo' );
+
+				testUtils.sinon.assert.calledOnce( spy );
+				testUtils.sinon.assert.calledWith( spy, options );
+			} );
 		} );
 	} );
 } );
