+++ packages/ckeditor5-core/tests/editor/utils/dataapimixin.js
@@ -5,73 +5,13 @@
 
 import DataApiMixin from '../../../src/editor/utils/dataapimixin.js';
 import Editor from '../../../src/editor/editor.js';
-import testUtils from '../../../tests/_utils/utils.js';
-import { getData, setData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model.js';
 
 describe( 'DataApiMixin', () => {
-	let editor;
+	it( 'it returns the passed parameter for compatibility reasons', () => {
+		class CustomEditor extends Editor {}
 
-	beforeEach( () => {
-		// eslint-disable-next-line new-cap
-		class CustomEditor extends DataApiMixin( Editor ) {}
+		const editor = new CustomEditor();
 
-		editor = new CustomEditor();
-		editor.model.document.createRoot( '$root', 'main' );
-		editor.model.document.createRoot( '$root', 'secondRoot' );
-		editor.model.schema.extend( '$text', { allowIn: '$root' } );
-		editor.fire( 'ready' ); // (#6139)
-	} );
-
-	afterEach( async () => {
-		await editor.destroy();
-	} );
-
-	it( 'is compatible with `mix` function', () => {
-		expect( DataApiMixin ).have.property( 'setData' ).to.be.a( 'function' );
-		expect( DataApiMixin ).have.property( 'getData' ).to.be.a( 'function' );
-	} );
-
-	describe( 'setData()', () => {
-		it( 'should be added to editor interface', () => {
-			expect( editor ).have.property( 'setData' ).to.be.a( 'function' );
-		} );
-
-		it( 'should set data of the first root', () => {
-			editor.setData( 'foo' );
-
-			expect( getData( editor.model, { rootName: 'main', withoutSelection: true } ) ).to.equal( 'foo' );
-		} );
-	} );
-
-	describe( 'getData()', () => {
-		testUtils.createSinonSandbox();
-
-		it( 'should be added to editor interface', () => {
-			expect( editor ).have.property( 'getData' ).to.be.a( 'function' );
-		} );
-
-		it( 'should get data of the first root', () => {
-			setData( editor.model, 'foo' );
-
-			expect( editor.getData() ).to.equal( 'foo' );
-		} );
-
-		it( 'should get data of the second root', () => {
-			setData( editor.model, 'bar', { rootName: 'secondRoot' } );
-
-			expect( editor.getData( { rootName: 'secondRoot' } ) ).to.equal( 'bar' );
-		} );
-
-		it( 'should pass options object to data.get() method internally', () => {
-			const spy = testUtils.sinon.spy( editor.data, 'get' );
-			const options = { rootName: 'main', trim: 'none' };
-
-			setData( editor.model, 'foo' );
-
-			expect( editor.getData( options ) ).to.equal( 'foo' );
-
-			testUtils.sinon.assert.calledOnce( spy );
-			testUtils.sinon.assert.calledWith( spy, options );
-		} );
+		expect( DataApiMixin( editor ) ).to.equal( editor );
 	} );
 } );
