+++ packages/ckeditor5-editor-classic/tests/classiceditor.js
@@ -52,11 +52,6 @@ describe( 'ClassicEditor', () => {
 			expect( editor.data.processor ).to.be.instanceof( HtmlDataProcessor );
 		} );
 
-		it( 'mixes DataApiMixin', () => {
-			expect( ClassicEditor.prototype ).have.property( 'setData' ).to.be.a( 'function' );
-			expect( ClassicEditor.prototype ).have.property( 'getData' ).to.be.a( 'function' );
-		} );
-
 		it( 'mixes ElementApiMixin', () => {
 			expect( ClassicEditor.prototype ).have.property( 'updateSourceElement' ).to.be.a( 'function' );
 		} );
