+++ packages/ckeditor5-editor-inline/src/inlineeditor.ts
@@ -10,7 +10,6 @@
 import {
 	Editor,
 	Context,
-	DataApiMixin,
 	ElementApiMixin,
 	attachToForm,
 	secureSourceElement,
@@ -49,7 +48,7 @@ import { isElement as _isElement } from 'lodash-es';
  * Read more about initializing the editor from source or as a build in
  * {@link module:editor-inline/inlineeditor~InlineEditor.create `InlineEditor.create()`}.
  */
-export default class InlineEditor extends DataApiMixin( ElementApiMixin( Editor ) ) {
+export default class InlineEditor extends ElementApiMixin( Editor ) {
 	/**
 	 * @inheritDoc
 	 */
