+++ packages/ckeditor5-ui/docs/_snippets/examples/bootstrap-ui-inner.js
@@ -13,7 +13,7 @@ import { CloudServices } from '@ckeditor/ckeditor5-cloud-services';
 import { CS_CONFIG } from '@ckeditor/ckeditor5-cloud-services/tests/_utils/cloud-services-config.js';
 // Interfaces to extend basic Editor API.
 // Helper function that binds editor with HTMLForm element.
-import { DataApiMixin, Editor, ElementApiMixin, attachToForm } from '@ckeditor/ckeditor5-core';
+import { Editor, ElementApiMixin, attachToForm } from '@ckeditor/ckeditor5-core';
 // The easy image integration.
 import { EasyImage } from '@ckeditor/ckeditor5-easy-image';
 import { Enter } from '@ckeditor/ckeditor5-enter';
@@ -26,10 +26,10 @@ import { EditorUI, EditorUIView, InlineEditableUIView } from '@ckeditor/ckeditor
 import { UndoEditing } from '@ckeditor/ckeditor5-undo';
 // Helper function for adding interfaces to the Editor class.
 // Helper function that gets data from HTML element that the Editor is attached to.
-import { ElementReplacer, getDataFromElement, mix } from '@ckeditor/ckeditor5-utils';
+import { ElementReplacer, getDataFromElement } from '@ckeditor/ckeditor5-utils';
 
 // Extending the Editor class, which brings base editor API.
-export default class BootstrapEditor extends Editor {
+export default class BootstrapEditor extends ElementApiMixin( Editor ) {
 	constructor( element, config ) {
 		super( config );
 
@@ -74,10 +74,6 @@ export default class BootstrapEditor extends Editor {
 		} );
 	}
 }
-
-// Mixing interfaces, which extends basic editor API.
-mix( BootstrapEditor, DataApiMixin );
-mix( BootstrapEditor, ElementApiMixin );
 
 // The class organizing the UI of the editor, binding it with existing Bootstrap elements in DOM.
 class BootstrapEditorUI extends EditorUI {
